+++ packages/core/src/components/TextBox.tsx
@@ -8,6 +8,9 @@
 
 import * as React from 'react';
 
+import { TextDirection, ThemeContext } from '../theme/ThemeContext';
+import { classNames } from '../utils/classNames';
+
 export const TextBox: React.FC<{
     ariaLabel?: string;
     placeholder?: string;
@@ -15,14 +18,22 @@ export const TextBox: React.FC<{
     value?: string;
     onChange: (value: string) => void;
     onKeyDown?: (e: React.KeyboardEvent) => void;
-}> = ({ ariaLabel = '', placeholder = '', type = 'text', value = '', onChange, onKeyDown = () => {} }) => (
-    <input
-        aria-label={ariaLabel}
-        className="rpv-core__textbox"
-        placeholder={placeholder}
-        type={type}
-        value={value}
-        onChange={(e: React.ChangeEvent<HTMLInputElement>) => onChange(e.target.value)}
-        onKeyDown={onKeyDown}
-    />
-);
+}> = ({ ariaLabel = '', placeholder = '', type = 'text', value = '', onChange, onKeyDown = () => {} }) => {
+    const { direction } = React.useContext(ThemeContext);
+    const isRtl = direction === TextDirection.RightToLeft;
+
+    return (
+        <input
+            aria-label={ariaLabel}
+            className={classNames({
+                'rpv-core__textbox': true,
+                'rpv-core__textbox--rtl': isRtl,
+            })}
+            placeholder={placeholder}
+            type={type}
+            value={value}
+            onChange={(e: React.ChangeEvent<HTMLInputElement>) => onChange(e.target.value)}
+            onKeyDown={onKeyDown}
+        />
+    );
+};