+++ packages/core/src/types/RenderViewer.ts
@@ -9,6 +9,7 @@
 import * as React from 'react';
 
 import { SpecialZoomLevel } from '../structs/SpecialZoomLevel';
+import type { ThemeContextProps } from '../theme/ThemeContext';
 import type { PdfJs } from '../types/PdfJs';
 import type { Slot } from './Slot';
 
@@ -19,6 +20,7 @@ export interface RenderViewer {
     pageWidth: number;
     rotation: number;
     slot: Slot;
+    themeContext: ThemeContextProps;
     openFile(file: File): void;
     jumpToPage(page: number): void;
     rotate(degree: number): void;